requirements.txt
iperf3
netifaces==0.10.4
pyyaml==3.12
pyzmq==16.0.2
msgpack-python==0.4.8
psutil==5.3.0
scapy==2.3.3
parse==1.8.2